Made FilePicker user FileBrowser. Have to still check FilePicker.tsx for redundancies. Also found a bug where if you refresh while in session, you'll try to collab with yourself
